requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
SQLAlchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
loguru==0.7.2
python-jose==3.3.0
bcrypt==4.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
